show delete files and refs

+++ components/forms/edit-video.tsx
@@ -135,25 +135,37 @@ export default function EditVideoForm({ videoInfo }: EditFormProps) {
       let targetThumbFile: File | null = null;
 
       try {
-        const thumbnail = await pickerRef.current?.generateThumbnail();
-        if (!thumbnail) throw new Error("Failed to generate thumbnail");
+        let thumbUri = videoInfo.thumbUri;
+        let thumbTimestamp = videoInfo.thumbTimestamp;
 
-        const rendered = await ImageManipulator.manipulate(thumbnail).renderAsync();
-        const { uri } = await rendered.saveAsync();
+        const currentPickerTimestamp = pickerRef.current?.getThumbTimestamp();
 
-        const thumbTimestamp = Math.round(thumbnail.requestedTime * 1000);
+        if (
+          currentPickerTimestamp !== undefined &&
+          currentPickerTimestamp !== videoInfo.thumbTimestamp
+        ) {
+          const thumbnail = await pickerRef.current?.generateThumbnail();
+          if (!thumbnail) throw new Error("Failed to generate thumbnail");
 
-        const fileId = createId();
-        targetThumbFile = new File(VIDEOS_DIR, `${fileId}_ppid_${videoInfo.id}.jpg`);
+          const rendered = await ImageManipulator.manipulate(thumbnail).renderAsync();
+          const { uri } = await rendered.saveAsync();
 
-        const tempThumbFile = new File(uri);
-        await tempThumbFile.move(targetThumbFile);
+          thumbTimestamp = Math.round(thumbnail.requestedTime * 1000);
+
+          const fileId = createId();
+          targetThumbFile = new File(VIDEOS_DIR, `${fileId}_ppid_${videoInfo.id}.jpg`);
+
+          const tempThumbFile = new File(uri);
+          await tempThumbFile.move(targetThumbFile);
+
+          thumbUri = targetThumbFile.uri;
+        }
 
         await updateVideo({
           id: videoInfo.id,
           values: {
             ...parsedValues,
-            thumbUri: targetThumbFile.uri,
+            thumbUri,
             thumbTimestamp,
             orientation: values.orientation.value,
             createdAt: values.createdAt.toISOString(),
@@ -161,9 +173,11 @@ export default function EditVideoForm({ videoInfo }: EditFormProps) {
           },
         });
 
-        const oldThumbFile = new File(videoInfo.thumbUri);
-        if (oldThumbFile.exists && oldThumbFile.uri !== targetThumbFile.uri) {
-          oldThumbFile.delete();
+        if (targetThumbFile?.exists) {
+          const oldThumbFile = new File(videoInfo.thumbUri);
+          if (oldThumbFile.exists && oldThumbFile.uri !== targetThumbFile.uri) {
+            oldThumbFile.delete();
+          }
         }
 
         toast.success(`${values.title} updated successfully.`);
